HBX-1549: Extract 'org.hibernate.tool.hbm2x.MetaAttributeHelper#copyMultiMap(MultiMap,MultiMap)' into new class 'org.hibernate.tool.internal.util.MultiMapUtil'
